Source config from core agent for ddprofiling
