+++ comp/otelcol/converter/impl/testdata/dd-core-cfg/all/api-section/config-result.yaml
@@ -23,9 +23,14 @@ extensions:
   zpages/user-defined:
     endpoint: "localhost:55679"
   ddflare/user-defined:
+  ddprofiling: 
+    api: 
+      key: ggggg77777
+      site: datadoghq.eu
+
 
 service:
-  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined]
+  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined, ddprofiling]
   pipelines:
     traces:
       receivers: [otlp]
