+++ comp/otelcol/converter/impl/testdata/dd-core-cfg/apikey/api-set-no-key/config.yaml
@@ -22,9 +22,12 @@ extensions:
   zpages/user-defined:
     endpoint: "localhost:55679"
   ddflare/user-defined:
+  ddprofiling:
+    api: 
+      site: datadoghq.eu
 
 service:
-  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined]
+  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined, ddprofiling]
   pipelines:
     traces:
       receivers: [otlp]
