+++ comp/otelcol/converter/impl/testdata/dd-core-cfg/apikey/multiple-dd-exporter/config-result.yaml
@@ -25,9 +25,18 @@ extensions:
   zpages/user-defined:
     endpoint: "localhost:55679"
   ddflare/user-defined:
+  ddprofiling/1:
+    api:
+      key: ggggg77777
+      site: datadoghq.com
+  ddprofiling/2:
+    api: 
+      key: ggggg77777
+      site: datadoghq.com
+
 
 service:
-  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined]
+  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined, ddprofiling/1, ddprofiling/2]
   pipelines:
     traces:
       receivers: [otlp]
