+++ comp/otelcol/converter/impl/testdata/dd-core-cfg/site/multiple-dd-exporter/config.yaml
@@ -25,9 +25,15 @@ extensions:
   zpages/user-defined:
     endpoint: "localhost:55679"
   ddflare/user-defined:
+  ddprofiling/1:
+    api: 
+      key: abcde12345
+  ddprofiling/2:
+    api: 
+      key: abcde12345
 
 service:
-  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined]
+  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined, ddprofiling/1, ddprofiling/2]
   pipelines:
     traces:
       receivers: [otlp]
