+++ comp/otelcol/converter/impl/testdata/dd-core-cfg/site/unset/config.yaml
@@ -23,9 +23,13 @@ extensions:
   zpages/user-defined:
     endpoint: "localhost:55679"
   ddflare/user-defined:
+  ddprofiling:
+    api:
+      key: abcde12345
+      site:
 
 service:
-  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined]
+  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined, ddprofiling]
   pipelines:
     traces:
       receivers: [otlp]
